Uploaded Code Project
This was a class assignment for CS 317 (Artificial Intelligence) in the Winter Term 2024 at Knox College. I built a “Markov Babbler” that uses Markov Chains to “babble” random sentences.

I trained the babbler on a large amount of text (usually called a corpus), and then have it generate random sentences that sound vaguely like they were written by the same author as the training set.